Change transaction meaning (requested amount is retrieve from the safe right away)

+++ back/secondary/perobobbot.data/perobobbot.data.domain/src/main/java/perobobbot/data/domain/base/SafeEntityBase.java
@@ -77,7 +77,7 @@ public class SafeEntityBase extends PersistentObjectWithUUID {
                    .build();
     }
 
-    public void addToAmount(int amountToAdd) {
+    public void addToAmount(long amountToAdd) {
         this.amount += amountToAdd;
     }
 }
